Turn CLI inside out
This makes the yargs parsing, which is synchronous, a simple function
that returns a data structure which can then be used to run the
CLI. This way, effection drives the control.

One thing is that there is a chicken and the egg problem with loading
the config because we don't know what the config file is until we
parse the options and we don't know the all the options until we parse
the config. This temporarily punts on this by removing the
`config-file` option. What we get in return is better validation of
the options because the choices and defaults are all drawn from the
config.