Actually run a test suite in CLI tests
It turns out that the `testFiles` property in .bigtest.json was not
actually not matching anything, because the pattern was looking for
`**/*.test.ts` but `passing.ts` didn't have the `.test` in its suffix.

This renames it to use the standard test suffix, and also adds an
ignore inside mocha for the fixtures directory so that those tests
won't be picked up. Also, I added some nice steps to the test suite so
that its output is a little less trivial.

An alternative would be to not use a test suffix at all, but there is
the possibility that we will have other files in the fixtures
directory that would then conflict.