readyResource for building ready state resources
A ton of resources require some initial phase of "boot up" before they are ready to be used. However, this is tricky in a resource based world, since we want to have whatever happens before ready actually survive. A readyResource is a resource which blocks until the `ready` function is called. This is just an abstraction to simplify this common pattern.