Make Agent resource more generally applicable
When writing the code that comprises the @bigtest/nodedriver, we will
have a total of three agents.

1. The web agent used to control web applications under test.
2. the unit test helper agent used to simulate an agent from within
test cases.
3. the node agents in @bigtest/nodedriver.

There is actually a lot of behavior that can be shared between these
three agents, and so this is a first step at consolidating the
elements of the agent that we know will be shared.

Since an agent _must_ send a connection message to register itself
with the orchestrator, this has been added to the `start()`
operation. Now, when that operation returns successfully, it will have
already sent the connection message.

Also in this PR, a subscribable is added over all of the comands to be
received from the orchestrator. E.g.

```ts
yield agent.commands.filter(command => command.type === 'run')
  .forEach(run => doRun());
```

The subscription will complete in the event that the socket is
closed.

Finally, since it is not a requirement that our sockets support the
`dispatchEvent` method of the `EventTarget` interface in order for our
API to listen to their messages, the API was narrowed to only require
`addEventListener` and `removeEventListener` since the underlying socket
libraries often don't implement those. We will have to do a bit of
gymnastics to get the typings to work out because the `on()` method
from `@effection/events` still requires the full `EventTarget` api,
but these can be removed once that type requirement [is also
relaxed][1]

[1]: https://github.com/thefrontside/effection/pull/149