Interactor is generic over its specification
Currently an interactor is typed `Interactor<E extends Element>`, this
changes the tpye to `Interactor<E extends Element, S extends
InteractorSpecification<E>>` that is the type of an interactor class and
its instance includes a reference to the type of the specification that
generated it.

The reason that this is useful is because it allows methods on the
interactor to refer to the specification. As part of #359, I would like
to add the ability to check for a given filter via `has({ filter:
whatever })`, but this requires knowledge of the type of the filter
specification to make it type-safe.

Why didn't we do this in the first place then? It does have some
downsides. The type of the specification is not easily namable, so when
an interactor needs to be referenced somewhere where its type must be
explicitly given, then that is not easily done. Also, previously we had
not yet worked out how to work around the variance issues that crop up
around `find`, so we were unable to make this work.