Annotate type of Socket.send()
Type declarations serve not only as correctness check, but also as an
educational tool. This is why the TypeScript compiler copies over
documentation comments on public interfaces into the declaration file.

Given that, Without any annotation, the type declaration file will show that
Socket.send is of type:

```ts
export declare class Socket {
  send(data: unknown): Generator<Promise<void>, any, any>
}
```

which doesn't really capture the intent. The Socket.send is actually
an `Operation` which does not yield any value, and should be yielded
to. When you see it in the docs, you want it to have it's primary
type, even though that type is an anlias.

By annonating the method explicitly, the type declaration is:

```ts
export declare class Socket {
  send(data: unknown): Operation<void>;
}
````