Harvest coverage metrics
One of the key use-cases is to be able to collect coverage data to
determine exactly how well your tests are doing, and what you might
need to do to improve the reliability of it. Because BigTest makes no
assumptions about how an app is the build, this part needs to be
handled by the actual app. That said, once there _is_ instrumentation,
BigTest can harvest the metrics for you and help you inject them into
your build pipeline at the appropriate point.

This PR is a first step in that process. It only handles harvesting of
coverage metrics from the agent and making sure that they are sent
back to to the orchestrator along with the `run:end` event. It does
not deal with integrating that coverage into the test run state, or
writing it out to a file from the cli.

To do this, the `__coverage__` object is checked after each lane run,
and if present, is merged into any prior coverage maps from
previous lanes. Once the test run is completed, the coverage data is
added to the `run:end`