Add interactive init command
This is a pretty basic init command, which only tackles the most basic configuration needed. It uses an interactive prompt created via the node `readline` package.

At first I attempted to do this using `ink`, but the results were not satisfying. Not only did it become much more complicated than it needed to be, but `ink` needs to emulate a proper terminal readline, which leads to very sub-par results. Not only do keybindings not work properly, but the resulting prompt also looks pretty janky.

If we do want to use `ink`, we can use it for those questions where it makes sense, but not for those that prompt for a string/number.

Closes #156