Report coverage from the CLI
We need [coverage reporting][1] in BigTest, and this is the front half
of the changes needed to implement it.

With coverage data [being harvested][2] from the web agents, we can
now take it and aggregate it on the server into a single coverage map
that spans coverage of the same application over multiple
browsers. This means that if you have code like the following:

```js
if (isFireFox()) {
  // some browser specific code.
}
```

The coverage metrics will change as you run begin to add different
browsers to your test suite, which is a very good thing.

Once the aggregate coverage data is integrated into the atom and
associated with the highest level `TestRunState`, it can be queried
over graphql.

On the client end, this adds a `coverage` field to the project which
has the output directory for the reports and the list of report
formats to output. That way when you pass the `--coverage` option to
the cli command, it knows what to do. For example, if your
bigtest.json looks like:

```json
{
  "coverage": {
    "reports": ["lcov", "html"],
    "directory": "./coverage"
  }
}
```

Then running the tests with the `--coverage` option

```
$ bigtest test --coverage
```

will render both the `lcov` and the `html` report to the `./coverage`
directory. Without specifically requesting coverage reports with the
`--coverage` flag, the coverage data will not be requested via graphql at all.

One thing that did significantly change is the timing of when an agent
run was considered complete. We had been "short circuiting" the run at
the moment all of the steps and assertions were reported either
passing, failing, or disregared. However, in order to make sure that
we get the coverage data we cannot exit early. Instead, we have to
explicitly wait for the `run:end` event to be sent by the agent since
that is the event bearing the raw coverage metrics. This shouldn't be
a problem, but it does pass control of "done-ness" of a test run over
to the agent from the orchestrator which is a significant shift. On
the whole this makes sense since the amount of metadata likely to be
passed back about a test run from the agent is only likely to grow,
and only the agent can determine when it is time to report that data.

Open Questions

* We could add configuration flags for the `reports` and `directory`
  fields of the config, but that felt like a bit much.

* I'm not super thrilled that the orchestrator and CLI are coupled
  directly to the instanbul coverage library since I don't know it to
  be a standard, and in order to generate coverage for different
  platforms, we'll need to make it generic. I feel like coverage is a
  pretty good use-case for having a plugin and it pretty much touches
  on every single thing where we need to have an extension point.
  - to stash cross-lane data on the agent
  - extend the run:end event to report metadata
  - to extend the orchestrator state to house the metadata
  - extend the run command to aggregate coverage data
  - extend the GraphQL schema to add the coverage field
  - extend CLI to add --coverage option
  - extend the project config to have a "coverage" section
  - add a custom reporter to handle test results and render reports to
  - report configuration errors and runtime errors to the user

If we can implement coverage and application logging events via a
plugin, we've got something really powerful.

[1]: https://github.com/thefrontside/bigtest/issues/446
[2]: https://github.com/thefrontside/bigtest/pull/554