Introduce Printer, an abstraction to simplify formatters
Currently, all of the code in `format-helpers` is a bit of a mess, because it needs to manually keep track of prefixes, since we work very heavily with indenting code. This results in a lot of manual jiggling to get everything to print correctly. Also, there are probably a bunch of edge cases that we currently don't handle very well, where newlines show up in unexpected places.

This introduces the concept of a printer, which can write to a writable stream (usually stdout), and has an internal prefix. When we print to the printer it will prefix all printed lines. This also makes it very easy to write recursive code, or to reuse the same code in different context with different prefixes.

The resulting output looks completely identical as far as I can tell.