Improve build error formatting
This is sort of a hybrid approach compared to what was discussed in #594. The `frame` part of the build error is included in the message, whereas the source file/line/column is added as a stack frame.

IMO the `frame` part doesn't really stand on its own. It's confusingly named and not formatted in a way which makes it easy to reuse. With this change we don't need to add this field to our GraphQL, which IMO feels better.