Add `focused` filter for asserting on focus state
When trying to implement the [todomvc test suite][1] using bigtest,
the very first assertion is that the todo input field has focus, which
made it immediately apparent that where focus is depending on various
actions is quite important for usability and accessibility. Questions
like "where is the focus when I open an aplication url?" or "where
does the focus go when I navigate to a particular route?" are critical
to be able to ask.

This adds a `focused` filter to all of the focusable elements (link,
button, checkbox, radio, textfield)

[1]: https://github.com/tastejs/todomvc/blob/master/tests/test.js