pre-munge ts-node path so that it works with exec()
Our exec() command uses shellwords to split command and
arguments. However, it has a path sepearator bug that fails to parse
absolute paths. This "pre-munges" the `ts-node` command into unix
format so that it can be un-munged by cross-spawn. A hack for sure,
but without an upstream fix on the `shellwords` package. That's where
we are.